+++ tests/disparity_graph.cpp
@@ -47,21 +47,21 @@ BOOST_AUTO_TEST_CASE(check_nodes_existence)
 
     for (ULONG row = 0; row < 2; ++row)
     {
-        BOOST_CHECK(node_exists(disparity_graph, {{row, 0}, 0}));
-        BOOST_CHECK(node_exists(disparity_graph, {{row, 1}, 0}));
-        BOOST_CHECK(node_exists(disparity_graph, {{row, 2}, 0}));
+        BOOST_CHECK(node_exists(disparity_graph, {{0, row}, 0}));
+        BOOST_CHECK(node_exists(disparity_graph, {{1, row}, 0}));
+        BOOST_CHECK(node_exists(disparity_graph, {{2, row}, 0}));
 
-        BOOST_CHECK(node_exists(disparity_graph, {{row, 0}, 1}));
-        BOOST_CHECK(node_exists(disparity_graph, {{row, 1}, 1}));
+        BOOST_CHECK(node_exists(disparity_graph, {{0, row}, 1}));
+        BOOST_CHECK(node_exists(disparity_graph, {{1, row}, 1}));
 
-        BOOST_CHECK(!node_exists(disparity_graph, {{row, 0}, 2}));
-        BOOST_CHECK(!node_exists(disparity_graph, {{row, 1}, 2}));
+        BOOST_CHECK(!node_exists(disparity_graph, {{0, row}, 2}));
+        BOOST_CHECK(!node_exists(disparity_graph, {{1, row}, 2}));
 
-        BOOST_CHECK(!node_exists(disparity_graph, {{row, 2}, 1}));
+        BOOST_CHECK(!node_exists(disparity_graph, {{2, row}, 1}));
     }
-    BOOST_CHECK(!node_exists(disparity_graph, {{2, 0}, 0}));
-    BOOST_CHECK(!node_exists(disparity_graph, {{0, 3}, 2}));
-    BOOST_CHECK(!node_exists(disparity_graph, {{2, 3}, 2}));
+    BOOST_CHECK(!node_exists(disparity_graph, {{0, 2}, 0}));
+    BOOST_CHECK(!node_exists(disparity_graph, {{3, 0}, 2}));
+    BOOST_CHECK(!node_exists(disparity_graph, {{3, 2}, 2}));
 }
 
 BOOST_AUTO_TEST_CASE(check_reparametrization_indexing)
@@ -82,20 +82,20 @@ BOOST_AUTO_TEST_CASE(check_reparametrization_indexing)
     struct DisparityGraph disparity_graph{image, image, 3};
 
     BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 0}, 0}, 0), 0);
-    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{1, 0}, 0}, 0), 1);
+    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 1}, 0}, 0), 1);
     BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 0}, 1}, 0), 2);
     BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 0}, 2}, 0), 4);
-    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{1, 0}, 2}, 0), 5);
+    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 1}, 2}, 0), 5);
     BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 0}, 0}, 1), 6);
-    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{1, 0}, 0}, 1), 7);
+    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 1}, 0}, 1), 7);
     BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 0}, 1}, 1), 8);
-    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{1, 0}, 1}, 1), 9);
+    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 1}, 1}, 1), 9);
     BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 0}, 2}, 1), 10);
-    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{1, 0}, 2}, 1), 11);
+    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 1}, 2}, 1), 11);
     BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 0}, 0}, 2), 12);
     BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 0}, 0}, 3), 18);
-    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{1, 0}, 2}, 3), 23);
-    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 1}, 0}, 0), 24);
+    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{0, 1}, 2}, 3), 23);
+    BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{1, 0}, 0}, 0), 24);
     BOOST_CHECK_EQUAL(reparametrization_index_fast(disparity_graph, {{1, 1}, 2}, 3), 47);
 }
 
@@ -127,22 +127,22 @@ BOOST_AUTO_TEST_CASE(check_neighborhood)
 
     struct DisparityGraph disparity_graph{left_image, right_image, 3};
 
-    BOOST_CHECK(neighborhood_exists(disparity_graph, {0, 0}, {0, 1}));
     BOOST_CHECK(neighborhood_exists(disparity_graph, {0, 0}, {1, 0}));
+    BOOST_CHECK(neighborhood_exists(disparity_graph, {0, 0}, {0, 1}));
 
-    BOOST_CHECK(neighborhood_exists(disparity_graph, {0, 1}, {0, 2}));
+    BOOST_CHECK(neighborhood_exists(disparity_graph, {1, 0}, {2, 0}));
 
-    BOOST_CHECK(neighborhood_exists(disparity_graph, {0, 1}, {0, 0}));
-    BOOST_CHECK(neighborhood_exists(disparity_graph, {0, 1}, {1, 1}));
+    BOOST_CHECK(neighborhood_exists(disparity_graph, {1, 0}, {0, 0}));
+    BOOST_CHECK(neighborhood_exists(disparity_graph, {1, 0}, {1, 1}));
 
     BOOST_CHECK(!neighborhood_exists(disparity_graph, {0, 0}, {0, 0}));
     BOOST_CHECK(!neighborhood_exists(disparity_graph, {0, 0}, {1, 1}));
-    BOOST_CHECK(!neighborhood_exists(disparity_graph, {0, 0}, {2, 0}));
     BOOST_CHECK(!neighborhood_exists(disparity_graph, {0, 0}, {0, 2}));
+    BOOST_CHECK(!neighborhood_exists(disparity_graph, {0, 0}, {2, 0}));
 
-    BOOST_CHECK(!neighborhood_exists(disparity_graph, {0, 1}, {0, 1}));
-    BOOST_CHECK(!neighborhood_exists(disparity_graph, {0, 1}, {1, 0}));
-    BOOST_CHECK(!neighborhood_exists(disparity_graph, {0, 1}, {1, 2}));
+    BOOST_CHECK(!neighborhood_exists(disparity_graph, {1, 0}, {1, 0}));
+    BOOST_CHECK(!neighborhood_exists(disparity_graph, {1, 0}, {0, 1}));
+    BOOST_CHECK(!neighborhood_exists(disparity_graph, {1, 0}, {2, 1}));
 }
 
 BOOST_AUTO_TEST_CASE(check_edges_existence)
@@ -173,17 +173,17 @@ BOOST_AUTO_TEST_CASE(check_edges_existence)
 
     struct DisparityGraph disparity_graph{left_image, right_image, 2};
 
-    BOOST_CHECK(edge_exists(disparity_graph, {{{0, 0}, 0}, {{0, 1}, 0}}));
-    BOOST_CHECK(edge_exists(disparity_graph, {{{0, 0}, 0}, {{0, 1}, 1}}));
+    BOOST_CHECK(edge_exists(disparity_graph, {{{0, 0}, 0}, {{1, 0}, 0}}));
+    BOOST_CHECK(edge_exists(disparity_graph, {{{0, 0}, 0}, {{1, 0}, 1}}));
 
-    BOOST_CHECK(edge_exists(disparity_graph, {{{0, 1}, 0}, {{0, 0}, 0}}));
-    BOOST_CHECK(edge_exists(disparity_graph, {{{0, 1}, 1}, {{0, 0}, 0}}));
+    BOOST_CHECK(edge_exists(disparity_graph, {{{1, 0}, 0}, {{0, 0}, 0}}));
+    BOOST_CHECK(edge_exists(disparity_graph, {{{1, 0}, 1}, {{0, 0}, 0}}));
 
-    BOOST_CHECK(!edge_exists(disparity_graph, {{{0, 1}, 2}, {{0, 1}, 1}}));
-    BOOST_CHECK(!edge_exists(disparity_graph, {{{0, 0}, 1}, {{0, 1}, 2}}));
+    BOOST_CHECK(!edge_exists(disparity_graph, {{{1, 0}, 2}, {{1, 0}, 1}}));
+    BOOST_CHECK(!edge_exists(disparity_graph, {{{0, 0}, 1}, {{1, 0}, 2}}));
 
-    BOOST_CHECK(!edge_exists(disparity_graph, {{{0, 0}, 2}, {{0, 1}, 0}}));
-    BOOST_CHECK(!edge_exists(disparity_graph, {{{0, 1}, 0}, {{0, 1}, 2}}));
+    BOOST_CHECK(!edge_exists(disparity_graph, {{{0, 0}, 2}, {{1, 0}, 0}}));
+    BOOST_CHECK(!edge_exists(disparity_graph, {{{1, 0}, 0}, {{1, 0}, 2}}));
 }
 
 BOOST_AUTO_TEST_CASE(check_continuity_constraint)
@@ -214,11 +214,11 @@ BOOST_AUTO_TEST_CASE(check_continuity_constraint)
 
     struct DisparityGraph disparity_graph{left_image, right_image, 3};
 
-    BOOST_CHECK(!edge_exists(disparity_graph, {{{0, 0}, 2}, {{0, 1}, 0}}));
-    BOOST_CHECK(!edge_exists(disparity_graph, {{{0, 1}, 0}, {{0, 0}, 2}}));
+    BOOST_CHECK(!edge_exists(disparity_graph, {{{0, 0}, 2}, {{1, 0}, 0}}));
+    BOOST_CHECK(!edge_exists(disparity_graph, {{{1, 0}, 0}, {{0, 0}, 2}}));
 
-    BOOST_CHECK(!edge_exists(disparity_graph, {{{1, 0}, 2}, {{1, 1}, 0}}));
-    BOOST_CHECK(!edge_exists(disparity_graph, {{{1, 1}, 0}, {{1, 0}, 2}}));
+    BOOST_CHECK(!edge_exists(disparity_graph, {{{0, 1}, 2}, {{1, 1}, 0}}));
+    BOOST_CHECK(!edge_exists(disparity_graph, {{{1, 1}, 0}, {{0, 1}, 2}}));
 }
 
 BOOST_AUTO_TEST_CASE(check_initial_edges_penalties)
@@ -250,34 +250,34 @@ BOOST_AUTO_TEST_CASE(check_initial_edges_penalties)
     struct DisparityGraph disparity_graph{left_image, right_image, 3};
 
     BOOST_CHECK_CLOSE(
-        edge_penalty(disparity_graph, {{{0, 0}, 0}, {{0, 1}, 0}}),
+        edge_penalty(disparity_graph, {{{0, 0}, 0}, {{1, 0}, 0}}),
         0.0,
         0.5
     );
     BOOST_CHECK_CLOSE(
-        edge_penalty(disparity_graph, {{{0, 0}, 0}, {{1, 0}, 1}}),
+        edge_penalty(disparity_graph, {{{0, 0}, 0}, {{0, 1}, 1}}),
         1.0,
         0.5
     );
 
     BOOST_CHECK_CLOSE(
-        edge_penalty(disparity_graph, {{{0, 1}, 1}, {{0, 2}, 0}}),
+        edge_penalty(disparity_graph, {{{1, 0}, 1}, {{2, 0}, 0}}),
         1.0,
         0.5
     );
 
     BOOST_CHECK_CLOSE(
-        edge_penalty(disparity_graph, {{{0, 1}, 0}, {{0, 0}, 2}}),
+        edge_penalty(disparity_graph, {{{1, 0}, 0}, {{0, 0}, 2}}),
         4.0,
         0.5
     );
     BOOST_CHECK_CLOSE(
-        edge_penalty(disparity_graph, {{{0, 1}, 2}, {{1, 0}, 0}}),
+        edge_penalty(disparity_graph, {{{1, 0}, 2}, {{0, 1}, 0}}),
         4.0,
         0.5
     );
     BOOST_CHECK_CLOSE(
-        edge_penalty(disparity_graph, {{{0, 1}, 2}, {{1, 1}, 2}}),
+        edge_penalty(disparity_graph, {{{1, 0}, 2}, {{1, 1}, 2}}),
         0.0,
         0.5
     );
@@ -315,19 +315,19 @@ BOOST_AUTO_TEST_CASE(check_initial_nodes_penalties)
     BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 0}, 1}), 0.0, 0.5);
     BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 0}, 2}), 1.0, 0.5);
 
-    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 1}, 0}), 1.0, 0.5);
-    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 1}, 1}), 4.0, 0.5);
-
-    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 2}, 0}), 0.0, 0.5);
-
-    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{1, 0}, 0}), 0.0, 0.5);
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{1, 0}, 0}), 1.0, 0.5);
     BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{1, 0}, 1}), 4.0, 0.5);
-    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{1, 0}, 2}), 1.0, 0.5);
+
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{2, 0}, 0}), 0.0, 0.5);
+
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 1}, 0}), 0.0, 0.5);
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 1}, 1}), 4.0, 0.5);
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 1}, 2}), 1.0, 0.5);
 
     BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{1, 1}, 0}), 0.0, 0.5);
     BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{1, 1}, 1}), 1.0, 0.5);
 
-    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{1, 2}, 0}), 0.0, 0.5);
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{2, 1}, 0}), 0.0, 0.5);
 }
 
 BOOST_AUTO_TEST_CASE(check_edges_penalties)
@@ -359,27 +359,27 @@ BOOST_AUTO_TEST_CASE(check_edges_penalties)
     struct DisparityGraph disparity_graph{left_image, right_image, 3};
 
     disparity_graph.reparametrization[
-        reparametrization_index(disparity_graph, {{0, 0}, 0}, {0, 1})
+        reparametrization_index(disparity_graph, {{0, 0}, 0}, {1, 0})
     ] = -2.0;
 
     BOOST_CHECK_CLOSE(
-        edge_penalty(disparity_graph, {{{0, 0}, 0}, {{0, 1}, 0}}),
+        edge_penalty(disparity_graph, {{{0, 0}, 0}, {{1, 0}, 0}}),
         -2.0,
         0.5
     );
     BOOST_CHECK_CLOSE(
-        edge_penalty(disparity_graph, {{{0, 0}, 0}, {{0, 1}, 1}}),
+        edge_penalty(disparity_graph, {{{0, 0}, 0}, {{1, 0}, 1}}),
         -1.0,
         0.5
     );
 
     BOOST_CHECK_CLOSE(
-        edge_penalty(disparity_graph, {{{0, 1}, 0}, {{0, 0}, 0}}),
+        edge_penalty(disparity_graph, {{{1, 0}, 0}, {{0, 0}, 0}}),
         -2.0,
         0.5
     );
     BOOST_CHECK_CLOSE(
-        edge_penalty(disparity_graph, {{{0, 1}, 0}, {{0, 0}, 1}}),
+        edge_penalty(disparity_graph, {{{1, 0}, 0}, {{0, 0}, 1}}),
         1.0,
         0.5
     );
@@ -414,13 +414,13 @@ BOOST_AUTO_TEST_CASE(check_nodes_penalties)
     struct DisparityGraph disparity_graph{left_image, right_image, 3};
 
     disparity_graph.reparametrization[
-        reparametrization_index(disparity_graph, {{0, 0}, 0}, {0, 1})
+        reparametrization_index(disparity_graph, {{0, 0}, 0}, {1, 0})
     ] = -2.0;
 
     BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 0}, 0}), 3.0, 0.5);
     BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 0}, 1}), 0.0, 0.5);
 
-    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 1}, 0}), 1.0, 0.5);
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{1, 0}, 0}), 1.0, 0.5);
 }
 
 BOOST_AUTO_TEST_SUITE_END()